Rollback: Add container to Tab component
