--- src/pages/Admin/Home/Dashboard/Menu/Tab/interfaces.ts
@@ -1,17 +1,0 @@
-import { FunctionComponent } from "react";
-import { ApprovableEntityType } from "../../../../../../interfaces/Approvable";
-
-export interface ITabContainerProps {
-  selected: boolean;
-  iconTitle: string;
-  onClick: () => void;
-  type: ApprovableEntityType;
-}
-
-export interface ITabProps extends Omit<ITabContainerProps, "type"> {
-  Icon: FunctionComponent<IIcon>;
-}
-
-interface IIcon {
-  className?: string;
-}
