+++ src/pages/Admin/Home/Dashboard/Menu/component.tsx
@@ -1,5 +1,7 @@
 import React, { FunctionComponent } from "react";
 import { Tab } from "./Tab";
+import { CompanyIcon } from "../CompanyIcon";
+import { ApplicantIcon } from "../ApplicantIcon";
 import { IMenuProps } from "./interfaces";
 import { APPLICANT, COMPANY } from "$typenames";
 import styles from "./styles.module.scss";
@@ -14,13 +16,13 @@ export const Menu: FunctionComponent<IMenuProps> = (
   <div className={styles.menuContent}>
     <Tab
       selected={filter.approvableEntityTypes.includes(COMPANY)}
-      type={COMPANY}
+      Icon={CompanyIcon}
       iconTitle={translations.companyIconTitle}
       onClick={() => addType(COMPANY)}
     />
     <Tab
       selected={filter.approvableEntityTypes.includes(APPLICANT)}
-      type={APPLICANT}
+      Icon={ApplicantIcon}
       iconTitle={translations.applicantIconTitle}
       onClick={() => addType(APPLICANT)}
     />
